Reorganize file heirarchy a bit more, rework render and anim somewhat, add some more code for parsing data from pfobj files

+++ src/render/vertex.h
@@ -4,7 +4,7 @@
 #include <GL/gl.h>
 
 struct joint_weight{
-    GLfloat joint_idx;
+    GLint   joint_idx;
     GLfloat weight;
 };
